refactor: extract parseReimbursements to own file for easier testing
