+++ src/common.js
@@ -1,0 +1,6 @@
+
+const baseUrl = 'https://extranet.malakoffmederic.com'
+
+module.exports = {
+  baseUrl
+}
