+++ src/index.js
@@ -12,7 +12,9 @@ const {
   errors,
   retry
 } = require('cozy-konnector-libs')
-const moment = require('moment')
+
+const { parseRemboursements, tokenify } = require('./parsers')
+const { baseUrl } = require('./common')
 
 let request = requestFactory()
 const j = request.jar()
@@ -22,7 +24,9 @@ request = requestFactory({
   json: false
 })
 
-const baseUrl = 'https://extranet.malakoffmederic.com'
+const getRequestOptions = () => ({
+  jar: j
+})
 
 module.exports = new BaseKonnector(start)
 
@@ -40,7 +44,7 @@ function start(fields) {
       })
     )
     .then(fetchRemboursements)
-    .then(parseRemboursements)
+    .then(res => parseRemboursements(res, getRequestOptions))
     .then(entries =>
       saveBills(entries, fields, {
         timeout: Date.now() + 60 * 1000,
@@ -123,102 +127,4 @@ function fetchRemboursements() {
   return request(`${baseUrl}/espaceClient/sante/tbs/redirectionAction.do`)
 }
 
-function parseRemboursements($) {
-  const result = []
-
-  // get the list of reimbursements rows
-  $('#tableauxRemboursements > .body > .toggle').each(function() {
-    const $header = $(this).find('.headerRemboursements')
-
-    const amount = convertAmount($header.find('.montant').text())
-    const date = moment($header.find('#datePaiement').val(), 'x')
-    const isThirdPartyPayer =
-      $(this)
-        .find('.dateEmission')
-        .text()
-        .indexOf('professionnels de santé') !== -1
-
-    // unique id for reimbursement
-    const idReimbursement = $header.find('#idDecompte').val()
-
-    let fileurl = $header.find('#tbsRembExportPdf').attr('href')
-    fileurl = `${baseUrl}${fileurl}`
-
-    const $subrows = $(this).find('> .body tbody tr')
-    let beneficiary = null
-    $subrows.each(function() {
-      const data = $(this)
-        .find('td, th')
-        .map(function() {
-          return $(this)
-            .text()
-            .trim()
-        })
-        .get()
-
-      if (data.length === 1) {
-        // we have a beneficiary line
-        beneficiary = data[0]
-      } else {
-        // a normal line with data
-        const originalAmount = convertAmount(data[data.length - 2])
-        const originalDate = moment(
-          $(this)
-            .find('#datePrestation')
-            .val(),
-          'x'
-        ).toDate()
-        const subtype = data[1]
-        // unique id for the prestation line. May be usefull
-        const idPrestation = $(this)
-          .find('#idPrestation')
-          .val()
-        const socialSecurityRefund = convertAmount(data[3])
-        result.push({
-          type: 'health_costs',
-          isThirdPartyPayer,
-          subtype,
-          vendor: 'Malakoff Mederic',
-          date: date.toDate(),
-          fileurl,
-          filename: getFileName(date, idReimbursement),
-          requestOptions: {
-            jar: j
-          },
-          amount,
-          idReimbursement,
-          idPrestation,
-          beneficiary,
-          socialSecurityRefund,
-          originalAmount,
-          originalDate,
-          isRefund: true
-        })
-      }
-    })
-  })
-
-  return result
-}
-
-function getFileName(date, idDecompte) {
-  // you can have multiple reimbursements for the same day
-  return `${date.format('YYYYMMDD')}_${idDecompte}_malakoff_mederic.pdf`
-}
-
-function convertAmount(amount) {
-  amount = amount.replace(' €', '').replace(',', '.')
-  return parseFloat(amount)
-}
-
-function tokenify(number) {
-  var tokenbuf = []
-  var charmap =
-    '1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ*$'
-  var remainder = number
-  while (remainder > 0) {
-    tokenbuf.push(charmap.charAt(remainder & 0x3f))
-    remainder = Math.floor(remainder / 64)
-  }
-  return tokenbuf.join('')
-}
+module.exports.parseRemboursements = parseRemboursements
